requirements.txt
Pillow==8.3.2
pytest==6.2.5